# Conflicts:
#	models/tt_dit/utils/sweep_mm_block_sizes.py

+++ ttnn/ttnn/operations/embedding.py
@@ -8,7 +8,8 @@ import ttnn
 def _golden_function(input_tensor: ttnn.Tensor, weight: ttnn.Tensor, **_):
     import torch
 
-    output_tensor = torch.nn.functional.embedding(input_tensor, weight)
+    # TTNN indices can be uint32, while torch embedding requires signed integer indices.
+    output_tensor = torch.nn.functional.embedding(input_tensor.to(torch.int64), weight)
     return output_tensor
